sync database's directory entries on start up
Badger isn't currently syncing the directory entries of the database
directories when it starts up. This could lead to the directories being
lost if the host crashes prior to flushing the directories to the disk.
Sync the directories on start up to ensure they are flushed to disk.
We sync the directories regardless whether we created them or not to
ensure the directory is flushed to disk if it was created previously
but not yet flushed.